Diverse pipolins reconstructed by ExplorePipolin
Predicted protein-coding genes are represented by arrows, indicating the direction of transcription and more common genes are colored following Prokka annota-tion (piPolB in red, tyrosine recombinase in brown, UDG in cyan, excisionase in purple and metallohydrolase in magenta). When detected, sequence gaps are indicated as black pointer, whereas the atts and tRNAs are represented as blue and green rectangle, respectively. The greyscale on the right reflects the percent of amino acid identity between pairs of sequences. The image was generated by EasyFig software using tBlastX for elements comparison. Names of the analized genomes are indicated on the left and colored by taxonomy: magenta, Alphaproteobacteria; red, Gammaproteobacteria; purple, fungi, forest green, Actinobacteria and green, Firmicutes.